Refactor file structure for tests
